requirements.txt
openai
# anthropic
phidata
python-dotenv
fastapi[standard]
